Syslog:TEST: Check that stderr is empty
A bug revealed that setting a source-interface plus a port generates a
'not defined' error even though the /etc/rsyslog.d/vyatta-log.conf file
is error free.

This commit adds support for checking vyatta_update_syslog.pl's stderr
channel in addition to the existing 'generated config' checks.
Essentially we test to see that the stderr log is empty.

Part of: VRVDR-41584
Fixes: VRVDR-48988