Make the syslog global target obsolete
The global target, which is basically a short cut for /var/log/message
file, is no longer the definitive source of system logs.

The journal is the definitive source of system logs, and 'show log' is
now based on the journal instead of /var/log/messages

'set system syslog file <name> ....' and 'show syslog file <name>' can
be used to direct messages to a file.

Closes: VRVDR-49099