update rsyslog config after static host config
In commit "b674e04 top level commit script for static host mapping"
a call to vyatta_update_syslog after static host configuration update was
removed. While this call appeared to be unneccessary on first look - on a
deeper look we find that the static host mapping may have an effect on
source address selection for rsyslog forwarding config. We don't need
to restart the rsyslog service as it is already done in
vyatta_update_syslog.pl.

Part of the work for VRVDR-49155