syslog: resolv static hosts targets in vrfs
In vrfs, If a resolv the syslog host target if the host is defined in
static host mappings, use the resolved ip address instead of hostname.

Also in vrfs use the target's ip address family to select the source
address from the source interface.

This is a rework from the older version that included this static
host mapping based resolutions in vrf. The reason for attempting
to resolve this hostnames in vrf is that rsyslogd omfwd doesn't have
the ability resolve DNS names in a vrf.

VRVDR-49322

(cherry picked from commit 415dc8dd9e623babd4cad0ffe6f97bf8e905d69e)