Syslog TCP and Source Interface cannot co-exist
The source-interface feature is currently only supported for UDP
in rsyslog. While there is an error message from rsyslog omfwd to
indicate that specifying a source address is not supported for TCP,
it is preferable to disallow source-interface and protocol TCP being
configured for syslog. If in the future TCP is supported upstream for
this feature, then the yang must statements can be removed, which is
permissible as it makes the yang less restrictive.

VRVDR-49401

(cherry picked from commit 94dc6a28fdc729910813f6c8b002232656efda9c)