Syslog: IPv6 source address search does not need to check for link-local
The search for IPv6 source address was previously based on operational
output and so a check was made to exclude link-local addresses. As the
addresses obtained with the current approach of checking statistics
never include link-locals, the check is superfluous and can be removed.

VRVDR-49365

(cherry picked from commit 76fad81627a64a250047de484ae7ca9f499fc845)