Syslog: Source interface should only be used if admin state is up
A requirement is for the source interface admin state to be up so as
to use an IP address configured on this as the source address. This
check was present prior to converting syslog handling to a module,
so now reintroduce it.

VRVDR-49365

(cherry picked from commit eba5ff57214924cd0c37f6106ea917dd04394a5b)