Allow customers to configure more NMIs and panics
A customer may desire well defined behavior during NMIs and other
events (such as OOM). Expose this via YANG so that we can use
deviations to set different default values on certain platforms.

Note that due to a bug in configd, we add a postconfig.d script to
set the initial values after the configuration is first loaded.

system options reboot-on-panic is now deprecated. This setting
was essentially always true (due to the defaults in the sysctl
.conf file) and only had an effect when set to false. If set
to false, this will take precedence over system fault-behavior
reboot-on-panic.

Bugfix: VRVDR-49991