SNMP syslog traps not sent for target in VRF
Issue: VRVDR-50130

The VRF list to send traps to was empty, and
hence no traps were sent to trap receivers in
VRF. The list was being obtained from the rsyslog
vrf directory - /run/rsyslog/vrf, which no longer
exists. This is not needed with the upstream VRF
changes. Fix uses the getvrflist command to fetch
the vrf list.