vyatta_update_login: Pass node status to update()
The $status variable passed to the update() subroutine of a given
login subsystem module typically indicates the status of the config
tree owned by that module, as returned by listNodeStatus().

However if the auth-chain configuration is enforcing an authentication
method other than one we call update() on, $status always has the
value "disabled" regardless of the actual config state.

TacplusServer.pm does not correctly apply its config when this occurs.
To help address this we pass an additional value to update() which is
always the value returned by listNodeStatus().

Other modules will ignore the new argument, but can be updated to
make use of it as required.

VRVDR-50552

(cherry picked from commit 798f507bd1bae8888271e38f346efa1f99cefe22)