Move YANG deviation file into yang-platform directory
Closes: VRVDR-50742

The vyatta-cfg-system repo has a Jenkins file that runs DRAM on each PR submission / update.  This uses the '-d' option meaning that we get an error due to duplicate inclusion of the SIAD deviations yang file.  To avoid this, and to follow (admittedly very recent!) best practice, the file is moved from yang/ to yang-platform/.

NB - the danos deviation file is NOT relocated as this is designed to be in ALL DANOS ISOs.  Arguably it should be *installed* to /usr/share/configd/yang on the vRouter, but be stored separately, but for now DRAM will pass so leaving well alone.