Add missing revision statement.
Closes: VRVDR-52079

When this file was ported to DANOS from VR project, a typedef that had been removed got reinstated by accident, with no corresponding revision statement update.  This is causing an exception in the nightly yang review for DANOS ISOs.

We can't remove the typedef as it has shipped in DANOS:Shipping:2005, albeit unused, so fix is simply to update the revision statement.