Default hostname has changed back to vyatta on recent master images
The default hostname for livecd images was changed from 'vyatta' to
'node' due to rebranding requirements for the 2005 release as per
VRVDR 50781. A recent change now results in top level yang defaults
being configured after boot as per VRVDR 51256, the problem being that
the yang default host-name had been left unchanged as 'vyatta'. The
solution is to remove this default, given that the default host-name
is managed elsewhere.

VRVDR-52495