remove the siad deviation for configuring NMIs and panics
The siad specific deviation for configuring the NMIs and panics is
no longer wanted. The siad platform should have the standard behaviour
for this, with access to the CLI to change it if required.

VRVDR-52568

(cherry picked from commit 9f05e6469e953d5d0b3eaeccf8f419ec9c934447)

Conflicts:
	debian/vyatta-system-misc-siad-deviations-v1-yang.install
	yang/vyatta-system-misc-siad-deviations-v1.yang