Allow script for ARP parameters to handle all parameters in a single call
The script for ARP and ND parameters needs to be called only once if
there is the resulting native configuration depends on more than one
vRouter configuration parameter. Ahead of this change for ARP, and
without any behavioral change for ND, convert the script so as to allow
both a single call for all parameters, or as currently a call per
parameter. Update the ND yang accordingly.

VRVDR-50363