Improve warning message when IPv6 is enabled and DAD is still failing
If IPv6 is operationally disabled due to DAD failure and strict-dad is
configured, then on manually enabling IPv6 again, an error "restore
<address> on <interface> failed (RTNETLINK answers: Permission denied)"
is shown for every address on the given interface. Display a single
error in this case and make it more helpful by indicating why the
addresses were not restored.

VRVDR-53629