RadiusServer.pm: correctly detect libpam-radius-auth ver.
Since 6e31db74e47c6924350977b283e21299ab1fc907 we were not correctly
varying the pam_radius_auth.cfg configuration file formatting based
on the installed version of libpam-radius-auth.

This resulted in us always writing a source IP value of 0 in the
configuration file. In versions of libpam-radius-auth prior to 1.4
(as we currently use), which don't support setting a source IP, this
resulted in us treating 0 as the VRF to bind to, effectively breaking
RADIUS authentication.

Since run3() always returns true regardless of the executed process'
exit code, we must check $? in order to choose the correct behaviour.

VRVDR-54038