Kernel: Generate ND packets with correct DSCP field
Use the ndisc_tclass kernel parameter to set the traffic class to
network control (CS6) for IPv6 ND packets originated by the kernel.
This is more efficient than using iptables and addresses the concern
that the DSCP field was previously not being set for ICMPv6 redirect
messages sent by the kernel.

VRVDR-54665