ARP stale time does not behave as one might expect for kernel entries
Configuring ARP stale-time sets the gc_stale_time kernel parameter used
to determine how often to check for stale entries in kernel's ARP cache.
If no ARP table-size is configured, this is the expiry time as expected.
If the table-size is configured though, the kernel garbage collector
does not remove entries if there are less than gc_thresh1 = size/8 of
them, and removes entries after 5s if there are more than gc_thresh2 =
size/2 of them. Update the description and help text to indicate that
the stale-time for the kernel ARP cache is conditional on behaviour of
the garbage collector.

VRVDR-54256