Fix issue of negative temperature causing show command traceback.
The "Sensor Reading" regexp matching for a number without a decimal point
did not allow a prior + or -. This meant that for a negative value, such
as a temperature, it would fail to match the regexp, and return "null"
as the value. This would cause the matching in the get-state YANG to fail
(as it expected an "int32" for the field), resulting in show command
"show hardware sensor" giving a python traceback.

VRVDR-56095

(cherry picked from commit 49dd80f6b2a8a9d1a8d684d7d768ae254df92675)